requirements.txt
Django==1.8.2
django-admin-bootstrapped==2.5.1
django-autocomplete-light==2.2.3
django-bootstrap3==5.4.0
django-crispy-forms==1.4.0
django-guardian==1.3
django-tables2==1.0.4
django-taggit==0.17.0
six==1.9.0
